+++ app/src/main/java/org/simple/clinic/AppDatabase.kt
@@ -17,6 +17,8 @@ import org.simple.clinic.home.overdue.OverdueAppointment
 import org.simple.clinic.home.overdue.OverdueAppointment_Old
 import org.simple.clinic.medicalhistory.Answer
 import org.simple.clinic.medicalhistory.MedicalHistory
+import org.simple.clinic.monthlyReports.Questionnaire
+import org.simple.clinic.monthlyReports.QuestionnaireType
 import org.simple.clinic.overdue.Appointment
 import org.simple.clinic.overdue.AppointmentCancelReason
 import org.simple.clinic.overdue.callresult.CallResult
@@ -92,12 +94,13 @@ import org.simple.clinic.drugs.search.Answer as DrugAnswer
       PatientFts::class,
       PatientPhoneNumberFts::class,
       BusinessIdFts::class,
-      PatientAddressFts::class
+      PatientAddressFts::class,
+      Questionnaire::class
     ],
     views = [
       PatientSearchResult::class
     ],
-    version = 105,
+    version = 106,
     exportSchema = true
 )
 @TypeConverters(
@@ -128,7 +131,8 @@ import org.simple.clinic.drugs.search.Answer as DrugAnswer
     DrugCategory.RoomTypeConverter::class,
     DrugAnswer.RoomTypeConverter::class,
     DrugFrequency.RoomTypeConverter::class,
-    Outcome.RoomTypeConverter::class
+    Outcome.RoomTypeConverter::class,
+    QuestionnaireType.RoomTypeConverter::class
 )
 abstract class AppDatabase : RoomDatabase() {