+++ app/src/main/java/org/simple/clinic/monthlyReports/Questionnaire.kt
@@ -1,0 +1,18 @@
+package org.simple.clinic.monthlyReports
+
+import android.os.Parcelable
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import kotlinx.parcelize.Parcelize
+import java.util.UUID
+
+@Entity(tableName = "Questionnaire")
+@Parcelize
+data class Questionnaire(
+    val uuid: UUID,
+
+    @PrimaryKey
+    val questionnaire_type: QuestionnaireType,
+
+    val layout: String,
+) : Parcelable